calculateScore() seems to be working just fine.
I need to change how I store words in wordList, so they can have a point value associated with them.
Again with the dictionary? At first glance that seems the correct way to do it.
This way you can see longest, as well as highest point value, words.

There should also be a check for words that use 7 tiles, as those are worth a bonus on the board as well.