Improved the buildWord() now called testWord() [testWord3() is the most recentand improved version].
scrabbleSolver now outputs the longest, and highest scoring word(s).

I'm not sure what else to improve or change with this script.
Maybe build a GUI for it, or allow for non-english and/or external dictionaries[it is
using the dictionary included in Mac OS X, 10.6.8] and letter sets.

Adding the traditional bonus or "bingo" for using all 7 tiles in a player's hand is a
simple addition.

Otherwise, I'm happy with what I wrote.

Learning about how to use dictionaries a bit better, taking base classes and extending
them, and thinking about optimization are all good things.

Woot!